+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.5
 * Remove unused avro property from `build.gradle.kts`
+* Trim schemaless data to only pass the fields that are in the table
 
 # 1.2.4
 * Adjusting underlying client version to 0.7.0